Custom way for storing authorization information
Until now the RhapsodySdkWrapper stores user authorization information
(access token, refresh token and catalog) by the Java Preferences
mechanism directly. This would fail on Android. Now you can pass an
AuthorizationStore implementation in the constructor of RhapsodySdkWrapper
to define how (if) the information is stored.
There is a PreferencesAuthorizationStore which was the default
implementation by now. Additionally there is a TransientAuthorizationStore
which doesn't save the information. An Android project can now implement
an SharedPreferencesAuthorizationStore for example.